cephfs admin: move rmFlags type to a file for flags handling
