+++ cephfs/admin/fsadmin.go
@@ -143,18 +143,3 @@ func modeString(m int, force bool) string {
 func uint64String(v uint64) string {
 	return strconv.FormatUint(uint64(v), 10)
 }
-
-type rmFlags struct {
-	force bool
-}
-
-func (f rmFlags) Update(m map[string]string) map[string]interface{} {
-	o := make(map[string]interface{})
-	for k, v := range m {
-		o[k] = v
-	}
-	if f.force {
-		o["force"] = true
-	}
-	return o
-}
